Refactor predictions and uncategorized pages to use server-side streaming with Suspense
- Split data fetching into separate server components for progressive loading
- Predictions page now shows layout immediately with loading indicators for each section
- Uncategorized page shows page structure while transactions load in background
- Maintains server-side fetching benefits while improving perceived performance
- Added PredictionChartSection, FutureChangesSection, InteractiveSimulationsSection
- Added TransactionsDataSection for uncategorized transactions
- Each section wrapped in individual Suspense boundaries for granular loading states
